pyproject.toml
[project]
name = "cluefin"
version = "0.1.0"
description = "Clearly Looking U Entered Financial Information"
readme = "README.md"
requires-python = ">=3.10"

[tool.uv.workspace]
members = [
    "packages/*",
    "apps/cluefin-cli",
]

[dependency-groups]
dev = [
    "coverage>=7.10.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=0.25.0",
    "python-dotenv>=1.1.1",
    "requests-mock>=1.12.1",
    "ruff>=0.12.3",
]

[tool.uv.sources]
cluefin-openapi = { workspace = true }
cluefin-cli = { workspace = true }

[tool.ruff]
line-length = 120
fix = true
target-version = "py311"

[tool.ruff.format]
docstring-code-format = true
[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501"]

[tool.pytest.ini_options]
markers = [
    "integration: integration tests",
    "slow: mark test as slow running",
]
addopts = "-ra"
testpaths = [
    "packages/cluefin-openapi/tests",
    "apps/cluefin-cli/tests",
]
